feat: - moved signout to its own page - basic styling and layout of login signup - added logout functionality
